+++ src/modules/achievements.js
@@ -3,86 +3,88 @@ const ACHIEVEMENTS = {
         caught: [
             {
                 name: 'First Catch',
-                value: 1
+                value: 1,
             },
             {
                 name: 'A Good Start',
-                value: 100
+                value: 100,
             },
             {
                 name: 'Thousand Club',
-                value: 1000
-            }
+                value: 1000,
+            },
         ],
         totalThrows: [
             {
                 name: 'Little league',
-                value: 100
+                value: 100,
             },
             {
                 name: 'Big league',
-                value: 10000
+                value: 10000,
             },
             {
                 name: 'Perfect aim',
-                value: 500000
+                value: 500000,
             },
             {
                 name: 'Check out these arm cannons',
-                value: 5000000
-            }
+                value: 5000000,
+            },
         ],
         totalPokeCoins: [
             {
                 name: 'Money Bags',
-                value: 1000000
-            }
+                value: 1000000,
+            },
         ],
         totalCatchCoins: [
             {
                 name: 'Catch Coin Bags',
-                value: 100000
-            }
+                value: 100000,
+            },
         ],
         totalBattleCoins: [
             {
                 name: 'Battle Coin Bags',
-                value: 100000
-            }
-        ]
+                value: 100000,
+            },
+        ],
     },
     dex: {
         caughtCount: [
             {
                 name: 'Kanto Collector',
-                value: 140
+                value: 140,
             },
             {
                 name: 'Johto Collector',
-                value: 240
-            }
+                value: 240,
+            },
         ],
         caught: [
             {
                 name: 'Professor Oak',
-                pokes: ['Squirtle', 'Bulbasaur', 'Charmander']
+                pokes: ['Squirtle', 'Bulbasaur', 'Charmander'],
             },
             {
                 name: 'Professor Elm',
-                pokes: ['Chikorita', 'Totodile', 'Cyndaquil']
+                pokes: ['Chikorita', 'Totodile', 'Cyndaquil'],
             },
             {
                 name: 'Professor Birch',
-                pokes: ['Treecko', 'Mudkip', 'Torchic']
+                pokes: ['Treecko', 'Mudkip', 'Torchic'],
             },
             {
                 name: 'Team Rocket',
-                pokes: ['Ekans', 'Koffing', 'Meowth']
+                pokes: ['Ekans', 'Koffing', 'Meowth'],
             },
             {
                 name: 'Mews',
-                pokes: ['Mew', 'Mewtwo']
-            }
-        ]
-    }
+                pokes: ['Mew', 'Mewtwo'],
+            },
+        ],
+    },
 };
+
+export default ACHIEVEMENTS;
