+++ src/modules/utilities.js
@@ -1,15 +1,17 @@
-function RNG(chance) {
+import POKEDEX from './db';
+
+export function RNG(chance) {
     const rnd = Math.random() * 100;
     return (rnd < chance);
 }
 
-function cloneJsonObject(object) {
+export function cloneJsonObject(object) {
     return JSON.parse(JSON.stringify(object));
 }
-function randomArrayElement(array) {
-    return array[Math.floor(Math.random() * array.length)]
+export function randomArrayElement(array) {
+    return array[Math.floor(Math.random() * array.length)];
 }
-function mergeArray(arr1, arr2) {
+export function mergeArray(arr1, arr2) {
     if (arr1.length === 0 && arr2.length > 0) {
         return arr2;
     }
@@ -21,17 +23,17 @@ function mergeArray(arr1, arr2) {
     }
     return [];
 }
-function isEmpty(obj) {
+export function isEmpty(obj) {
     return Object.keys(obj).length === 0 && obj.constructor === Object;
 }
-function camalCaseToString(text) {
-    let result = text.replace( /([A-Z])/g, " $1" );
+export function camelCaseToString(text) {
+    const result = text.replace(/([A-Z])/g, ' $1');
     return result.charAt(0).toUpperCase() + result.slice(1);
 }
-function flash(element) {
-    let op = 1;  // initial opacity
+export function flash(element) {
+    let op = 1; // initial opacity
     let fadeOut = true;
-    let timer = setInterval(function () {
+    const timer = setInterval(() => {
         if (op <= 0.1) {
             fadeOut = false;
         }
@@ -39,7 +41,7 @@ function flash(element) {
             clearInterval(timer);
         }
         element.style.opacity = op;
-        element.style.filter = 'alpha(opacity=' + op * 100 + ")";
+        element.style.filter = `alpha(opacity=${op * 100})`;
         if (fadeOut) {
             op -= op * 0.1;
         } else {
@@ -48,4 +50,7 @@ function flash(element) {
     }, 10);
 }
 
-const $ = (cssQuery) => document.querySelector(cssQuery);
+export const pokeById = (id) => POKEDEX[id - 1];
+export const pokeByName = (name) => POKEDEX.filter((el) => el.pokemon[0].Pokemon === name)[0];
+
+export const $ = (cssQuery) => document.querySelector(cssQuery);
